backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Document parsing (offline)
python-docx==1.1.2
PyMuPDF==1.24.14
unstructured[docx,pdf]>=0.15.0
markdown==3.7

# LLM integrations
httpx==0.27.2
openai==1.54.4
anthropic==0.40.0
google-generativeai==0.8.3

# Utilities
aiofiles==24.1.0
python-magic==0.4.27
chardet==5.2.0

# БД и авторизация (история анализов, JWT)
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
